\encoding{latin1}
\name{binda-package}
\alias{binda-package}
\docType{package}
\title{The binda Package}

\description{
The "binda" package implements functions for multi-class discriminant
analysis using binary predictors, for corresponding variable selection, 
and for dichotomizing continuous data.
}

\author{Sebastian Gibb and Korbinian Strimmer (\url{https://strimmerlab.github.io/})}
\references{

Gibb, S., and K. Strimmer. 2015.  Differential protein expression and peak selection in mass spectrometry data by binary discriminant analysis. Bioinformatics 31:3156-3162.
<DOI:10.1093/bioinformatics/btv334>
  
  Website: \url{https://strimmerlab.github.io/software/binda/}
}
\keyword{multivariate}
\seealso{
\code{\link{binda}}, \code{\link{binda.ranking}}, \code{\link{dichotomize}}, \code{\link{chances}}.
}
