\encoding{latin1}
\name{binda-package}
\alias{binda-package}
\docType{package}
\title{The binda Package}

\description{
The "binda" package implements functions for multi-class discriminant
analysis using binary predictors, for corresponding variable selection, 
and for dichotomizing continuous data.
}

\author{Sebastian Gibb and Korbinian Strimmer (\url{http://strimmerlab.org/})}
\references{

Gibb, S., and K. Strimmer. 2015.  Differential protein expression and peak selection in mass spectrometry data by binary discriminant analysis. Bioinformatics, to appear.
\url{http://arxiv.org/abs/1502.07959}

  
  Website: \url{http://strimmerlab.org/software/binda/}
}
\keyword{multivariate}
\seealso{
\code{\link{binda}}, \code{\link{binda.ranking}}, \code{\link{dichotomize}}, \code{\link{chances}}.
}
